@@ -3,33 +3,27 @@
 Object.defineProperty(exports, "__esModule", {
   value: true
 });
-Object.defineProperty(exports, "MetaTagsContext", {
-  enumerable: true,
-  get: function get() {
-    return _meta_tags_context.default;
-  }
-});
 Object.defineProperty(exports, "MetaTags", {
   enumerable: true,
   get: function get() {
-    return _meta_tags.default;
+    return _meta_tags["default"];
+  }
+});
+Object.defineProperty(exports, "MetaTagsContext", {
+  enumerable: true,
+  get: function get() {
+    return _meta_tags_context["default"];
   }
 });
 Object.defineProperty(exports, "ReactTitle", {
   enumerable: true,
   get: function get() {
-    return _react_title.default;
+    return _react_title["default"];
   }
 });
-exports.default = void 0;
-
+exports["default"] = void 0;
 var _meta_tags_context = _interopRequireDefault(require("./meta_tags_context"));
-
 var _meta_tags = _interopRequireDefault(require("./meta_tags"));
-
 var _react_title = _interopRequireDefault(require("./react_title"));
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
-
-var _default = _meta_tags.default;
-exports.default = _default;
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { "default": obj }; }
+var _default = exports["default"] = _meta_tags["default"];
