@@ -3,52 +3,35 @@
 Object.defineProperty(exports, "__esModule", {
   value: true
 });
-exports.default = void 0;
-
+exports["default"] = void 0;
 var _react = _interopRequireWildcard(require("react"));
-
 var _reactDom = _interopRequireDefault(require("react-dom"));
-
 var _utils = require("./utils");
-
 var _meta_tags_context = require("./meta_tags_context");
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
-
-function _interopRequireWildcard(obj) { if (obj && obj.__esModule) { return obj; } else { var newObj = {}; if (obj != null) { for (var key in obj) { if (Object.prototype.hasOwnProperty.call(obj, key)) { var desc = Object.defineProperty && Object.getOwnPropertyDescriptor ? Object.getOwnPropertyDescriptor(obj, key) : {}; if (desc.get || desc.set) { Object.defineProperty(newObj, key, desc); } else { newObj[key] = obj[key]; } } } } newObj.default = obj; return newObj; } }
-
-function _typeof(obj) { if (typeof Symbol === "function" && typeof Symbol.iterator === "symbol") { _typeof = function _typeof(obj) { return typeof obj; }; } else { _typeof = function _typeof(obj) { return obj && typeof Symbol === "function" && obj.constructor === Symbol && obj !== Symbol.prototype ? "symbol" : typeof obj; }; } return _typeof(obj); }
-
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { "default": obj }; }
+function _getRequireWildcardCache(e) { if ("function" != typeof WeakMap) return null; var r = new WeakMap(), t = new WeakMap(); return (_getRequireWildcardCache = function _getRequireWildcardCache(e) { return e ? t : r; })(e); }
+function _interopRequireWildcard(e, r) { if (!r && e && e.__esModule) return e; if (null === e || "object" != _typeof(e) && "function" != typeof e) return { "default": e }; var t = _getRequireWildcardCache(r); if (t && t.has(e)) return t.get(e); var n = { __proto__: null }, a = Object.defineProperty && Object.getOwnPropertyDescriptor; for (var u in e) if ("default" !== u && Object.prototype.hasOwnProperty.call(e, u)) { var i = a ? Object.getOwnPropertyDescriptor(e, u) : null; i && (i.get || i.set) ? Object.defineProperty(n, u, i) : n[u] = e[u]; } return n["default"] = e, t && t.set(e, n), n; }
+function _typeof(o) { "@babel/helpers - typeof"; return _typeof = "function" == typeof Symbol && "symbol" == typeof Symbol.iterator ? function (o) { return typeof o; } : function (o) { return o && "function" == typeof Symbol && o.constructor === Symbol && o !== Symbol.prototype ? "symbol" : typeof o; }, _typeof(o); }
 function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
-
-function _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }
-
-function _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); return Constructor; }
-
-function _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === "object" || typeof call === "function")) { return call; } return _assertThisInitialized(self); }
-
+function _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, _toPropertyKey(descriptor.key), descriptor); } }
+function _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, "prototype", { writable: false }); return Constructor; }
+function _callSuper(t, o, e) { return o = _getPrototypeOf(o), _possibleConstructorReturn(t, _isNativeReflectConstruct() ? Reflect.construct(o, e || [], _getPrototypeOf(t).constructor) : o.apply(t, e)); }
+function _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === "object" || typeof call === "function")) { return call; } else if (call !== void 0) { throw new TypeError("Derived constructors may only return object or undefined"); } return _assertThisInitialized(self); }
 function _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return self; }
-
-function _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }
-
-function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); if (superClass) _setPrototypeOf(subClass, superClass); }
-
-function _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf || function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }
-
-function _defineProperty(obj, key, value) { if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }
-
+function _isNativeReflectConstruct() { try { var t = !Boolean.prototype.valueOf.call(Reflect.construct(Boolean, [], function () {})); } catch (t) {} return (_isNativeReflectConstruct = function _isNativeReflectConstruct() { return !!t; })(); }
+function _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf.bind() : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }
+function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); Object.defineProperty(subClass, "prototype", { writable: false }); if (superClass) _setPrototypeOf(subClass, superClass); }
+function _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf ? Object.setPrototypeOf.bind() : function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }
+function _defineProperty(obj, key, value) { key = _toPropertyKey(key); if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }
+function _toPropertyKey(t) { var i = _toPrimitive(t, "string"); return "symbol" == _typeof(i) ? i : String(i); }
+function _toPrimitive(t, r) { if ("object" != _typeof(t) || !t) return t; var e = t[Symbol.toPrimitive]; if (void 0 !== e) { var i = e.call(t, r || "default"); if ("object" != _typeof(i)) return i; throw new TypeError("@@toPrimitive must return a primitive value."); } return ("string" === r ? String : Number)(t); }
 /** An wrapper component to wrap element which need to shifted to head **/
-var MetaTags =
-/*#__PURE__*/
-function (_Component) {
+var MetaTags = /*#__PURE__*/function (_Component) {
   _inherits(MetaTags, _Component);
-
   function MetaTags() {
     _classCallCheck(this, MetaTags);
-
-    return _possibleConstructorReturn(this, _getPrototypeOf(MetaTags).apply(this, arguments));
+    return _callSuper(this, MetaTags, arguments);
   }
-
   _createClass(MetaTags, [{
     key: "componentDidMount",
     value: function componentDidMount() {
@@ -66,7 +49,7 @@ function (_Component) {
     key: "componentWillUnmount",
     value: function componentWillUnmount() {
       if (this.temporaryElement) {
-        _reactDom.default.unmountComponentAtNode(this.temporaryElement);
+        _reactDom["default"].unmountComponentAtNode(this.temporaryElement);
       }
     }
   }, {
@@ -74,11 +57,9 @@ function (_Component) {
     value: function extractChildren() {
       var extract = this.context.extract;
       var children = this.props.children;
-
       if (!children) {
         return;
       }
-
       if (extract) {
         extract(children);
       }
@@ -87,48 +68,45 @@ function (_Component) {
     key: "handleChildrens",
     value: function handleChildrens() {
       var _this = this;
-
       var children = this.props.children;
-
       if (this.context.extract || !children) {
         return;
       }
-
-      var headComponent = _react.default.createElement("div", {
+      var headComponent = /*#__PURE__*/_react["default"].createElement("div", {
         className: "react-head-temp"
       }, children);
+      var root = _reactDom["default"].createRoot(this.temporaryElement);
+      root.render(function () {
+        var childStr = _this.temporaryElement.innerHTML;
 
-      _reactDom.default.render(headComponent, this.temporaryElement, function () {
-        var childStr = _this.temporaryElement.innerHTML; //if html is not changed return
-
+        //if html is not changed return
         if (_this.lastChildStr === childStr) {
           return;
         }
-
         _this.lastChildStr = childStr;
+        var tempHead = _this.temporaryElement.querySelector('.react-head-temp');
 
-        var tempHead = _this.temporaryElement.querySelector('.react-head-temp'); // .react-head-temp might not exist when triggered from async action
-
-
+        // .react-head-temp might not exist when triggered from async action
         if (tempHead === null) {
           return;
         }
-
         var childNodes = Array.prototype.slice.call(tempHead.children);
         var head = document.head;
-        var headHtml = head.innerHTML; //filter children remove if children has not been changed
+        var headHtml = head.innerHTML;
 
+        //filter children remove if children has not been changed
         childNodes = childNodes.filter(function (child) {
           return headHtml.indexOf(child.outerHTML) === -1;
-        }); //create clone of childNodes
+        });
 
+        //create clone of childNodes
         childNodes = childNodes.map(function (child) {
           return child.cloneNode(true);
-        }); //remove duplicate title and meta from head
+        });
 
+        //remove duplicate title and meta from head
         childNodes.forEach(function (child) {
           var tag = child.tagName.toLowerCase();
-
           if (tag === 'title') {
             var title = (0, _utils.getDuplicateTitle)();
             if (title) (0, _utils.removeChild)(head, title);
@@ -145,6 +123,7 @@ function (_Component) {
           }
         });
         (0, _utils.appendChild)(document.head, childNodes);
+        return headComponent;
       });
     }
   }, {
@@ -154,12 +133,8 @@ function (_Component) {
       return null;
     }
   }]);
-
   return MetaTags;
 }(_react.Component);
-
 _defineProperty(MetaTags, "contextType", _meta_tags_context.MetaContext);
-
-var _default = MetaTags;
-exports.default = _default;
+var _default = exports["default"] = MetaTags;
 module.exports = exports.default;
